refactor: Replace TeamNavigation with ParticipantNavigation in AssessmentPage
